Update index.html
Instead of the pretty and pink Negroni website, I chose to create an exercise website since I could list more. I used the first <li> as a placeholder for the homepage and used Spotify as a link that opens a different website, above the footer. The list so far doesn't contain working links, but it is a start!